Create IFS.cpp
If you want the program to recursively search for files without the iterative deepening approach (i.e., a standard depth-first search that goes as deep as possible in each path before backtracking), you can modify the IDDFS function to call DLS only once with a large depth value. This will effectively turn the search into a standard recursive depth-first search.

Here's the modified version of the FileSearchIDDFS.cpp program to do a straightforward recursive search:

Removed the IDDFS function and added a recursiveSearch function.
This function calls DLS with a MAX_DEPTH set to the maximum value for an int. This effectively removes the iterative deepening aspect and causes the program to perform a standard recursive search.
The main function is updated to reflect these changes.
You can compile and run this program in the same way as before, but now it will perform a simple recursive search starting from the specified directory.